+++ iocore/net/Makefile.am
@@ -154,6 +154,7 @@ libinknet_a_SOURCES = \
 	P_SSLUtils.h \
 	P_SSLClientCoordinator.h \
 	P_SSLClientUtils.h \
+	P_TLSKeyLogger.h \
 	P_OCSPStapling.h \
 	P_UDPConnection.h \
 	P_UDPIOEvent.h \
@@ -189,6 +190,7 @@ libinknet_a_SOURCES = \
 	OCSPStapling.cc \
 	TLSBasicSupport.cc \
 	TLSEarlyDataSupport.cc \
+	TLSKeyLogger.cc \
 	TLSSessionResumptionSupport.cc \
 	TLSSNISupport.cc \
 	TLSTunnelSupport.cc \
