+++ iocore/net/quic/QUICConfig.cc
@@ -28,7 +28,7 @@
 #include <records/I_RecHttp.h>
 
 #include "P_SSLConfig.h"
-#include "P_SSLUtils.h"
+#include "P_TLSKeyLogger.h"
 
 #include "QUICGlobals.h"
 #include "QUICTransportParameters.h"
